work on make orders

+++ contracts/Token.sol
@@ -4,24 +4,32 @@ pragma solidity ^0.8.17;
 contract Token {
     string public name;
     string public symbol;
-    uint public decimals = 18;
-    uint public totalSupply;
+    uint256 public decimals = 18;
+    uint256 public totalSupply;
 
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
-    event Transfer(address indexed _from, address indexed _to, uint _value);
-    event Approval(address indexed _owner, address indexed _spender, uint _value);
+    event Transfer(address indexed _from, address indexed _to, uint256 _value);
+    event Approval(
+        address indexed _owner,
+        address indexed _spender,
+        uint256 _value
+    );
 
-    constructor(string memory _name, string memory _symbol, uint _totalSupply ) {
+    constructor(
+        string memory _name,
+        string memory _symbol,
+        uint256 _totalSupply
+    ) {
         name = _name;
         symbol = _symbol;
         totalSupply = _totalSupply * (10**decimals);
         balanceOf[msg.sender] = totalSupply;
     }
 
-    modifier requireValue(uint _value) {
-        require(balanceOf[msg.sender] >= _value, 'not enough balance');
+    modifier requireValue(uint256 _value) {
+        require(balanceOf[msg.sender] >= _value, "not enough balance");
         _;
     }
 
@@ -30,19 +38,33 @@ contract Token {
         _;
     }
 
-    function transfer(address _to, uint _value) public requireValue(_value) requireAddress0(_to) returns(bool success) {
+    function transfer(address _to, uint256 _value)
+        public
+        requireValue(_value)
+        requireAddress0(_to)
+        returns (bool success)
+    {
         _transfer(msg.sender, _to, _value);
         return true;
     }
 
-    function _transfer(address _from, address _to, uint _value) internal {
+    function _transfer(
+        address _from,
+        address _to,
+        uint256 _value
+    ) internal {
         balanceOf[_from] = balanceOf[_from] - _value;
         balanceOf[_to] = balanceOf[_to] + _value;
 
         emit Transfer(_from, _to, _value);
     }
 
-    function approve(address _spender, uint _value) public requireValue(_value) requireAddress0(_spender) returns(bool success) {
+    function approve(address _spender, uint256 _value)
+        public
+        requireValue(_value)
+        requireAddress0(_spender)
+        returns (bool success)
+    {
         allowance[msg.sender][_spender] = _value;
 
         emit Approval(msg.sender, _spender, _value);
@@ -50,9 +72,13 @@ contract Token {
         return true;
     }
 
-    function transferFrom(address _from, address _to, uint _value) public returns(bool success) {
-        require(_value <= balanceOf[_from], 'not enough balance');
-        require(_value <= allowance[_from][msg.sender], 'not enough allowance');
+    function transferFrom(
+        address _from,
+        address _to,
+        uint256 _value
+    ) public returns (bool success) {
+        require(_value <= balanceOf[_from], "not enough balance");
+        require(_value <= allowance[_from][msg.sender], "not enough allowance");
 
         allowance[_from][msg.sender] = allowance[_from][msg.sender] - _value;
